Encrypt sqlite data on disk with key stored in Keychain.
When signing in initially a random passphrase is generated that is
stored in the Keychain together with the session data.
This prevents cached data to be extracted from disk without having
access to the Keychain item.
