+++ Mactrix/Models/UserSession.swift
@@ -1,0 +1,62 @@
+import AsyncAlgorithms
+import Foundation
+import KeychainAccess
+import MatrixRustSDK
+import OSLog
+import Security
+import SwiftUI
+import UI
+import UniformTypeIdentifiers
+import Utils
+
+struct UserSession: Codable {
+    let accessToken: String
+    let refreshToken: String?
+    let userID: String
+    let deviceID: String
+    let homeserverURL: String
+    let oidcData: String?
+    let storeID: String
+    let storePassphrase: String
+
+    init(session: Session, storeID: String, storePassphrase: String) {
+        accessToken = session.accessToken
+        refreshToken = session.refreshToken
+        userID = session.userId
+        deviceID = session.deviceId
+        homeserverURL = session.homeserverUrl
+        oidcData = session.oidcData
+        self.storeID = storeID
+        self.storePassphrase = storePassphrase
+    }
+
+    var session: Session {
+        Session(accessToken: accessToken,
+                refreshToken: refreshToken,
+                userId: userID,
+                deviceId: deviceID,
+                homeserverUrl: homeserverURL,
+                oidcData: oidcData,
+                slidingSyncVersion: .native)
+    }
+
+    fileprivate static var keychainKey: String { "UserSession" }
+
+    func saveUserToKeychain() throws {
+        let keychainData = try JSONEncoder().encode(self)
+        let keychain = Keychain(service: applicationID)
+        try keychain.set(keychainData, key: Self.keychainKey)
+    }
+
+    static func loadUserFromKeychain() throws -> Self? {
+        Logger.matrixClient.debug("Load user from keychain")
+        /* #if DEBUG
+             if true {
+                 return try JSONDecoder().decode(Self.self, from: DevSecrets.matrixSession.data(using: .utf8)!)
+             }
+         #endif */
+        let keychain = Keychain(service: applicationID)
+        guard let keychainData = try keychain.getData(keychainKey) else { return nil }
+        return try JSONDecoder().decode(Self.self, from: keychainData)
+    }
+}
